Replace f32 usage with Decimal (#110)
Decimal is now used to retain decimal accuracy. As per the requirements of the OCPP specification.

+++ src/v2_0_1/datatypes/sampled_value_type.rs
@@ -1,3 +1,5 @@
+use rust_decimal::Decimal;
+
 use super::signed_meter_value_type::SignedMeterValueType;
 use super::unit_of_measure_type::UnitOfMeasureType;
 use crate::v2_0_1::enumerations::location_enum_type::LocationEnumType;
@@ -11,7 +13,8 @@ use crate::v2_0_1::enumerations::reading_context_enum_type::ReadingContextEnumTy
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct SampledValueType {
-    pub value: f32,
+    #[serde(with = "rust_decimal::serde::arbitrary_precision")]
+    pub value: Decimal,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub context: Option<ReadingContextEnumType>,
     #[serde(skip_serializing_if = "Option::is_none")]
